remove legacy contact info
