+++ gossip/src/crds_gossip_pull.rs
@@ -665,7 +665,6 @@ pub(crate) mod tests {
         super::*,
         crate::{
             crds_data::{CrdsData, Vote},
-            legacy_contact_info::LegacyContactInfo,
             protocol::Protocol,
         },
         itertools::Itertools,
@@ -1560,14 +1559,8 @@ pub(crate) mod tests {
             node
         };
         {
-            let caller = CrdsValue::new(CrdsData::from(&node), &keypair);
+            let caller: CrdsValue = CrdsValue::new(CrdsData::from(&node), &keypair);
             assert_eq!(get_max_bloom_filter_bytes(&caller), 1175);
-            verify_get_max_bloom_filter_bytes(&mut rng, &caller, num_items);
-        }
-        let node = LegacyContactInfo::try_from(&node).unwrap();
-        {
-            let caller = CrdsValue::new(CrdsData::LegacyContactInfo(node), &keypair);
-            assert_eq!(get_max_bloom_filter_bytes(&caller), 1136);
             verify_get_max_bloom_filter_bytes(&mut rng, &caller, num_items);
         }
         let node = {
@@ -1580,12 +1573,6 @@ pub(crate) mod tests {
         {
             let caller = CrdsValue::new(CrdsData::from(&node), &keypair);
             assert_eq!(get_max_bloom_filter_bytes(&caller), 1155);
-            verify_get_max_bloom_filter_bytes(&mut rng, &caller, num_items);
-        }
-        let node = LegacyContactInfo::try_from(&node).unwrap();
-        {
-            let caller = CrdsValue::new(CrdsData::LegacyContactInfo(node), &keypair);
-            assert_eq!(get_max_bloom_filter_bytes(&caller), 992);
             verify_get_max_bloom_filter_bytes(&mut rng, &caller, num_items);
         }
     }
