+++ gossip/src/legacy_contact_info.rs
@@ -1,12 +1,11 @@
 #[cfg(test)]
-use crate::contact_info::{sanitize_socket, ContactInfo, Error, Protocol, SOCKET_ADDR_UNSPECIFIED};
+use crate::{socketaddr, socketaddr_any};
 use {
     crate::crds_data::{reject_deserialize, MAX_WALLCLOCK},
     serde::Serialize,
     solana_pubkey::Pubkey,
     solana_sanitize::{Sanitize, SanitizeError},
-    solana_streamer::socket::SocketAddrSpace,
-    std::net::{IpAddr, SocketAddr},
+    std::net::SocketAddr,
 };
 
 /// Structure representing a node on the network
@@ -50,45 +49,6 @@ impl Sanitize for LegacyContactInfo {
     }
 }
 
-macro_rules! get_socket {
-    ($name:ident) => {
-        #[cfg(test)]
-        pub(crate) fn $name(&self) -> Option<SocketAddr> {
-            let socket = self.$name;
-            sanitize_socket(&socket).ok()?;
-            Some(socket)
-        }
-    };
-    ($name:ident, $quic:ident) => {
-        #[cfg(test)]
-        pub(crate) fn $name(&self, protocol: Protocol) -> Option<SocketAddr> {
-            let socket = match protocol {
-                Protocol::QUIC => self.$quic,
-                Protocol::UDP => self.$name,
-            };
-            sanitize_socket(&socket).ok()?;
-            Some(socket)
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! socketaddr {
-    ($ip:expr, $port:expr) => {
-        std::net::SocketAddr::from((std::net::Ipv4Addr::from($ip), $port))
-    };
-    ($str:expr) => {{
-        $str.parse::<std::net::SocketAddr>().unwrap()
-    }};
-}
-
-#[macro_export]
-macro_rules! socketaddr_any {
-    () => {
-        socketaddr!(std::net::Ipv4Addr::UNSPECIFIED, 0)
-    };
-}
-
 #[cfg(test)]
 impl Default for LegacyContactInfo {
     fn default() -> Self {
@@ -120,113 +80,11 @@ impl LegacyContactInfo {
     pub(crate) fn wallclock(&self) -> u64 {
         self.wallclock
     }
-
-    #[inline]
-    pub(crate) fn shred_version(&self) -> u16 {
-        self.shred_version
-    }
-
-    pub(crate) fn gossip(&self) -> Option<SocketAddr> {
-        let socket = self.gossip;
-        crate::contact_info::sanitize_socket(&socket).ok()?;
-        Some(socket)
-    }
-
-    get_socket!(tvu, tvu_quic);
-    get_socket!(tpu);
-    get_socket!(tpu_forwards);
-    get_socket!(tpu_vote);
-    get_socket!(rpc);
-    get_socket!(rpc_pubsub);
-    get_socket!(serve_repair, serve_repair_quic);
-
-    fn is_valid_ip(addr: IpAddr) -> bool {
-        !(addr.is_unspecified() || addr.is_multicast())
-        // || (addr.is_loopback() && !cfg_test))
-        // TODO: boot loopback in production networks
-    }
-
-    /// port must not be 0
-    /// ip must be specified and not multicast
-    /// loopback ip is only allowed in tests
-    // TODO: Replace this entirely with streamer SocketAddrSpace.
-    pub(crate) fn is_valid_address(addr: &SocketAddr, socket_addr_space: &SocketAddrSpace) -> bool {
-        addr.port() != 0u16 && Self::is_valid_ip(addr.ip()) && socket_addr_space.check(addr)
-    }
-}
-
-#[cfg(test)]
-impl TryFrom<&ContactInfo> for LegacyContactInfo {
-    type Error = Error;
-
-    fn try_from(node: &ContactInfo) -> Result<Self, Self::Error> {
-        macro_rules! unwrap_socket {
-            ($name:ident) => {
-                node.$name().unwrap_or(SOCKET_ADDR_UNSPECIFIED)
-            };
-            ($name:ident, $protocol:expr) => {
-                node.$name($protocol).unwrap_or(SOCKET_ADDR_UNSPECIFIED)
-            };
-        }
-        Ok(Self {
-            id: *node.pubkey(),
-            gossip: unwrap_socket!(gossip),
-            tvu: unwrap_socket!(tvu, Protocol::UDP),
-            tvu_quic: unwrap_socket!(tvu, Protocol::QUIC),
-            serve_repair_quic: unwrap_socket!(serve_repair, Protocol::QUIC),
-            tpu: unwrap_socket!(tpu, Protocol::UDP),
-            tpu_forwards: unwrap_socket!(tpu_forwards, Protocol::UDP),
-            tpu_vote: unwrap_socket!(tpu_vote, Protocol::UDP),
-            rpc: unwrap_socket!(rpc),
-            rpc_pubsub: unwrap_socket!(rpc_pubsub),
-            serve_repair: unwrap_socket!(serve_repair, Protocol::UDP),
-            wallclock: node.wallclock(),
-            shred_version: node.shred_version(),
-        })
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use {super::*, std::net::Ipv4Addr};
-
-    #[test]
-    fn test_is_valid_address() {
-        let bad_address_port = socketaddr!(Ipv4Addr::LOCALHOST, 0);
-        assert!(!LegacyContactInfo::is_valid_address(
-            &bad_address_port,
-            &SocketAddrSpace::Unspecified
-        ));
-        let bad_address_unspecified = socketaddr!(Ipv4Addr::UNSPECIFIED, 1234);
-        assert!(!LegacyContactInfo::is_valid_address(
-            &bad_address_unspecified,
-            &SocketAddrSpace::Unspecified
-        ));
-        let bad_address_multicast = socketaddr!([224, 254, 0, 0], 1234);
-        assert!(!LegacyContactInfo::is_valid_address(
-            &bad_address_multicast,
-            &SocketAddrSpace::Unspecified
-        ));
-        let loopback = socketaddr!(Ipv4Addr::LOCALHOST, 1234);
-        assert!(LegacyContactInfo::is_valid_address(
-            &loopback,
-            &SocketAddrSpace::Unspecified
-        ));
-        //        assert!(!LegacyContactInfo::is_valid_ip_internal(loopback.ip(), false));
-    }
-
-    #[test]
-    fn test_default() {
-        let ci = LegacyContactInfo::default();
-        assert!(ci.gossip.ip().is_unspecified());
-        assert!(ci.tvu.ip().is_unspecified());
-        assert!(ci.tpu_forwards.ip().is_unspecified());
-        assert!(ci.rpc.ip().is_unspecified());
-        assert!(ci.rpc_pubsub.ip().is_unspecified());
-        assert!(ci.tpu.ip().is_unspecified());
-        assert!(ci.tpu_vote.ip().is_unspecified());
-        assert!(ci.serve_repair.ip().is_unspecified());
-    }
+    use super::*;
 
     #[test]
     fn test_sanitize() {
